+++ mmodel/__init__.py
@@ -1,6 +1,6 @@
 from mmodel.graph import ModelGraph
 from mmodel.handler import MemHandler, PlainHandler, H5Handler
 from mmodel.model import Model
-from mmodel.modifier import basic_loop, zip_loop
+from mmodel.modifier import loop_modifier, zip_loop_modifier
 from mmodel.utility import subgraph_by_parameters, subgraph_by_nodes, modify_subgraph
 from mmodel.draw import draw_graph, draw_plain_graph