+++ tests/test_modifier.py
@@ -1,4 +1,4 @@
-from mmodel.modifier import basic_loop, zip_loop
+from mmodel.modifier import loop_modifier, zip_loop_modifier
 import pytest
 
 @pytest.fixture
@@ -12,13 +12,13 @@ def mockedfunc():
 def test_basic_loop(mockedfunc):
     """Test redirect edges based on subgraph and subgraph node"""
 
-    loop_mod = basic_loop("b")
+    loop_mod = loop_modifier("b")
     looped = loop_mod(mockedfunc)
 
     assert looped(a=1, b=[1, 2, 3], c=4) == [6, 7, 8]
 
     # test the method with default
-    # loop_mod = basic_loop("c")
+    # loop_mod = loop_modifier("c")
     # looped_default = loop_mod(mock_func)
 
     # assert inspect.signature(looped_default).parameters['c'].default == [2]
@@ -29,13 +29,13 @@ def test_basic_loop(mockedfunc):
 def test_zip_loop(mockedfunc):
     """Test redirect edges based on subgraph and subgraph node"""
 
-    loop_mod = zip_loop(["a", "b"])
+    loop_mod = zip_loop_modifier(["a", "b"])
     looped = loop_mod(mockedfunc)
 
     assert looped(a=[0.1, 0.2, 0.3], b=[1, 2, 3], c=10) == [11.1, 12.2, 13.3]
 
     # test the method with default
-    # loop_mod = basic_loop("c")
+    # loop_mod = loop_modifier("c")
     # looped_default = loop_mod(mock_func)
 
     # assert inspect.signature(looped_default).parameters['c'].default == [2]